Consolidate into existing TestCommon
