+++ src/test/java/org/datadog/jmxfetch/TestCommon.java
@@ -7,10 +7,15 @@ import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.when;
 
 import com.beust.jcommander.JCommander;
+
+import java.io.BufferedWriter;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
+import java.io.FileWriter;
 import java.io.IOException;
 import java.lang.management.ManagementFactory;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashSet;
@@ -30,6 +35,30 @@ import org.datadog.jmxfetch.util.CustomLogger;
 import org.datadog.jmxfetch.util.LogLevel;
 import org.junit.After;
 import org.junit.BeforeClass;
+
+final class ConfigUtil {
+    public static Path writeConfigYamlToTemp(String content, String yamlName) throws IOException {
+        Path tempDirectory = Files.createTempDirectory("temp-dir");
+        // Create a temporary file within the temporary directory
+        Path tempFile = Files.createTempFile(tempDirectory, yamlName, ".yaml");
+
+        // Write the contents of the file to the temporary file
+        BufferedWriter writer = new BufferedWriter(new FileWriter(tempFile.toFile()));
+        writer.write(content);
+        writer.close();
+
+        return tempFile;
+    }
+
+    public static String concatWithNewlines(String... lines) {
+        StringBuilder sb = new StringBuilder();
+        for (String line : lines) {
+            sb.append(line).append(System.lineSeparator());
+        }
+        return sb.toString();
+    }
+}
+
 
 public class TestCommon {
     AppConfig appConfig = spy(AppConfig.builder().build());
@@ -137,6 +166,37 @@ public class TestCommon {
     protected void initApplication(String yamlFileName) throws FileNotFoundException, IOException {
         initApplication(yamlFileName, "");
     }
+
+    /*
+     * Init JMXFetch with the given YAML configuration template 
+     * The configuration can be specified as a yaml literal with each arg
+     * representing one line of the Yaml file
+     * Does not support any SD/AD features.
+    */
+    protected void initApplicationWithYamlLines(String... yamlLines)
+            throws IOException {
+        String yamlConfig = ConfigUtil.concatWithNewlines(yamlLines);
+        Path tempFile = ConfigUtil.writeConfigYamlToTemp(yamlConfig, "config");
+
+        String confdDirectory = tempFile.getParent().toString();
+        String yamlFileName = tempFile.getFileName().toString();
+        
+        List<String> params = new ArrayList<String>();
+        params.add("--reporter");
+        params.add("console");
+
+        if (confdDirectory != null) {
+            params.add("-c");
+            params.add(yamlFileName);
+            params.add("--conf_directory");
+            params.add(confdDirectory);
+            params.add("collect");
+        }
+        new JCommander(appConfig, params.toArray(new String[params.size()]));
+        this.app = new App(appConfig);
+        app.init(false);
+    }
+
 
     /** Run a JMXFetch iteration. */
     protected void run() {
